Change handling of '__all__' parsing when no elements could be parsed
Parsing of `__all__` an element can fail because the element is not
a literal (too complex for pydoctor) or because the element is not
of type `str` (almost certainly broken code). We attempt to parse
elements separately and will use every name we can parse.

The old implementation would ignore the assignment altogether if
no names could be parsed, while the new implementation doesn't
make that distinction and will assign the empty list to `mod.all`
if no names could be parsed.

It's hard to say which behavior is better, but at least the new
behavior is simpler.