Add test case that finds a reparented interface
The code that searches for a reparented interface at its new location
had a deceptively high coverage, since the the 'not found' case hits
every line. But the `if iface is None` statement never branched to
the False path, because we didn't actually have a test yet that
reparented an interface.