Refactor code that finds a renamed interface into a method of System
The problem solved by this code is not unique to interfaces, so we
might need it for other objects as well.

Usually I prefer to generalize code when the second use case is
discovered instead of when it is only assumed to exist, but in this
case splitting the code off into a separate method is already useful
as it improves the readability of _handle_implemented().