Split 'docs' tox env into one that just builds and one that tests too
While the existing 'docs' environment is very useful for CI, it does
work and logs data that are not relevant when you are writing new
documentation and just want to create an updated render.

I also added the `-W` option to `sphinx-build` for the CI version
('testdocs'), so our build will fail if there are warnings. All
warnings I saw from Sphinx so far have been serious; if that is
not true in general we might have to drop this option again.