Return None for nested calls in funcNameFromCall()
The original implementation of this method would call itself
recursively if the `func` attribute of a `Call` node is another
`Call` node. When switching from `astor` to `node2fullname()`,
I kept that behavior but replaced the recursive call by a `while`
loop.

However, when trying to write a test case to cover the `while`
body, I couldn't think of any plausible scenario in which this
feature is useful. It's easy enough to trigger: `f()()` will
do it, but none of the code using `funcNameFromCall()` would
want to treat a call to the returned object the same as a call
to the function.