Recognize custom property using '@mod.func' decoration
Only `ast.Name` nodes were checked until now, which meant that `@func`
would work but `@mod.func` would not, since the latter is parsed to an
`ast.Attribute` node.

We already have a way to handle both `Name` and `Attribute` nodes in
a single flow: `node2dottedname()`, so use that.