Support for google-style inline attribute docstrings.
Need to change how the parse_docstring() function was discovered in order to allow to
use a different parsing function depending on
the documentable type. This is currently only used by the google-style parser.
It could be used to create a new parser named 'autodetect' that would try it's best to
detect the docstring format and return the right parsing function on the fly. But this is something for later.