Remove FieldDesc.kind attribute
This attribute only contained actual information for unknown fields.
For every known field type, the kind was implied by the list in which
the FieldDesc objects are stored.

Instead of storing the unknown field name in FieldDesc, we now store
it in FieldHandler.unknowns as keys. This also replaces a conversion
step from FieldHandler.format() where we'd group the unknown fields
by field name before presentation.