Select the same monospace font for all code and code-like text
We had a different CSS rule for definition lines, which could lead
to different fonts being selected. And on my Linux desktop that
actually happened.

For other elements we had the same rule copy-pasted several times,
which gave the right result but is not great for maintenance, so
I combined those as well.