Add query string parameter for showing private API
Links and bookmarks can now opt to show private API on page load.

If a browser does not support searchParams, we fall back to the
current behavior of hiding private API on page load.

If a browser does not support the History API, we don't update the
location bar, but the toggle still works.