When fragment points to private API, auto-enable "show private"
The Intersphinx inventories that pydoctor produces include private
names. When the inventory or a naming scheme is used to link to
private API programmatically, the URL probably won't include the
`private` argument. That would lead to the linked item being
invisible.

A piece of JavaScript now looks up the target of the fragment and
if it's inside a node with the CSS class `private`, the private
API won't be hidden on page load.