Make Package a subclass of Module
Previously pydoctor had a Package object for the namespace and a Module
object for the code from the `__init__.py` source file. However, in the
Python Language Reference, it is explicitly stated that packages are
modules.

Having the Package class inherit from Module simplifies some code while
it complicates other code, but I think overall the simplifications
outweigh the complications.