Fix positioning of anchors using CSS tweaks
One problem was introduced between 20.12.x and 21.2.x when I changed
the page header to be a flexbox: this somehow makes browsers treat
the margin on the `h1` differently.

Another problem was an inconsistency between how Chrome and Firefox
render the page: Chrome considered the whitespace-only anchor to
have line height 0, while Firefox applied the font's line height.
By setting the font size to 0, both browsers use a line height of 0.