Insert project name + optional URL using a slot
There used to be a renderer for a possibly hyperlinked name and a slot
that would insert the name without a link. Now we use a slot at the
page level, to make it available everywhere.

In practice, this means the project name in the footer is now a link.