Remove "part" slot
This was already unused by our templates, but kept around to not break
Twisted. However, Twisted's templates will have to be redesigned
anyway when the template rework lands, so now is a good time to
remove this.

A comment states that the CSS needs to be removed as well, but I can't
find any CSS relevant to this, so I guess it was already removed as
part of earlier cleanups.