Update never-called deprecated functions in templatewriter.util
The `templatefilepath()` function was only called by pydoctor itself,
so it can be safely removed now.

The `templatefile()` function is not called by Twisted, but is
monkeypatched by Twisted. It does need to exist since Twisted will
save the original function and restore it after it's done. I updated
the docstring to make it clear that Twisted does not call it.