Fix three bugs and add tests for two of them.
The first one in how epytext raises an exception when parsing. Now it will fall back to plaintext thanks to this exception.

The second bug is regarding the node tree generation from epytext: It now exposes a function, set_nodes_parent() and re-uses this function to reparent the contents of the second level to the root level and avoid unecessary section in node tree AND keep .parent property coherent.

The last bug, is regarding validation of CSS indentifiers in the _PydoctorHTMLTranslator. Now there is an function that remove invalid characters to create valid CSS identifiers.