Use Dict[str, Documentable] as annotation for Documentable.contents.
Adjust the rest of the annotation in the codebase.

Add assertion when needed, like before reparent() call to assert the parent is a module.

Fix a potentially problematic None value in the Documentable.kind property in format_undocumented().

Change the unwrap() function in the test_astbuilder module to accept Optional[ParsedDocstring] but actually raises if the parsed_docstring is not an ParsedEpytextDocstring. This avoids a lot of extra cast() call or assertions in test_astbuilder.