Use ._alias_to attribute to store the aliases indirection instead of computing it from AST all the time.
Consider variables that are not on the root level as aliases, too. And warns only when they are actually overriden.

Move node2fullname to astutils module.

Delete overridenInCount from ClassPage since it was unused.

Re-export names that are not part of the current system with an alias.

Cleanup expandName() and associated.

Use directly taglink() in format_alias_value() when possible.

Add Documentable.aliases property. This is somewhat working in the tests but not always in real life.

Rename the redirected_from parameter to "indirections".

Speaking of tests, this commit also adds A LOT of new tests for the expandName() method.