Fix supported docformats missing from --help
Previously running pydoctor with -h or --help showed:

  --docformat=FORMAT    Format used for parsing docstrings. Supported values:
  --template-dir=PATH   Directory containing custom HTML templates. Can
                        repeat.

Note that the supported values for --docformat didn't show up.

This was caused by get_supported_docformats() returning an iterator,
which was then depleted with choices=list(_docformat_choices)
leaving nothing for the subsequent ', '.join(_docformat_choices).

The fix is simply to convert the iterator to a list right away.