Fix random module processing order
Previously modules to be processed were stored in a set.
A set however is an unordered collection (and adding items
can completely change the order depending on the object hash).

Since the Module class did not implement __hash__ every time
there was more than one module to be processed the order
in which they were processed was effectively random leading
to very hard to reproduce bugs.

Fixes #482 and #429.