Get rid of the AnnotationLinker - drop the verbose messages when an annotation is ambiguous (pydoctor is not a checker).
Get rid of ParsedDocstring.with_linker() this API did not make sens and was rather an anti-pattern.

Replace that with a new parameter of colorize_pyval(is_annotation: bool). obj_reference nodes can now be created with the attribute 'is_annotation' in which case they will be resolved a little bit differently.